requirements.txt
asgiref==3.7.2
asttokens==2.2.1
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
chart-studio==1.1.0
colorama==0.4.6
colorlover==0.3.0
comm==0.1.2
contourpy==1.0.6
cufflinks==0.17.3
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
Django==4.2.2
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
env==0.1.0
executing==1.2.0
fonttools==4.38.0
idna==3.4
ipykernel==6.20.2
ipython==8.9.0
ipywidgets==8.0.4
jedi==0.18.2
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
matplotlib==3.6.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.1
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
plotly==5.13.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.6
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
PyMuPDF==1.22.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
requests==2.28.1
requests-oauthlib==1.3.1
retrying==1.3.4
seaborn==0.12.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.6.2
tenacity==8.1.0
tornado==6.2
traitlets==5.9.0
tweepy==4.12.1
tzdata==2022.7
urllib3==1.26.13
wcwidth==0.2.6
widgetsnbextension==4.0.5
